INC-539: Extract analytics data mappers into utils file
